refactor and reorganize field line viz functions
