+++ synthesizAR/extrapolate/helpers.py
@@ -1,17 +1,124 @@
 """
 Helper routines for field extrapolation routines and dealing with vector field data
 """
+import warnings
+
 import numpy as np
+from scipy.interpolate import RegularGridInterpolator
 import astropy.time
 import astropy.units as u
 from astropy.coordinates import SkyCoord
 import astropy.constants as const
 import yt
 import sunpy.coordinates
-from sunpy.util.metadata import MetaDict
-from sunpy.map import GenericMap
+from sunpy.map import GenericMap, make_fitswcs_header
 
-__all__ = ['synthetic_magnetogram', 'magnetic_field_to_yt_dataset', 'from_local', 'to_local']
+__all__ = ['from_pfsspack', 'semi_circular_loop', 'synthetic_magnetogram',
+           'magnetic_field_to_yt_dataset', 'from_local', 'to_local']
+
+
+def from_pfsspack(pfss_fieldlines):
+    """
+    Convert fieldline coordinates output from the SSW package `pfss <http://www.lmsal.com/~derosa/pfsspack/>`_ 
+    into `~astropy.coordinates.SkyCoord` objects.
+
+    Parameters
+    ----------
+    pfss_fieldlines : `~numpy.recarray`
+        Structure produced by reading pfss output with `~scipy.io.readsav`
+
+    Returns
+    -------
+    fieldlines : `list`
+        Each entry is a `tuple` containing a `~astropy.coordinates.SkyCoord` object and a
+        `~astropy.units.Quantity` object listing the coordinates and field strength along the loop.
+    """
+    # Fieldline coordinates
+    num_fieldlines = pfss_fieldlines['ptr'].shape[0]
+    # Use HGC frame if possible
+    try:
+        frame = sunpy.coordinates.HeliographicCarrington(
+            obstime=sunpy.time.parse_time(pfss_fieldlines['now'].decode('utf-8')))
+    except ValueError:
+        warnings.warn('Assuming HGS frame because no date available for HGC frame')
+        frame = sunpy.coordinates.HeliographicStonyhurst()
+    fieldlines = []
+    for i in range(num_fieldlines):
+        # NOTE: For an unknown reason, there are a number of invalid points for each line output
+        # by pfss
+        n_valid = pfss_fieldlines['nstep'][i]
+        lon = (pfss_fieldlines['ptph'][i, :] * u.radian).to(u.deg)[:n_valid]
+        lat = 90 * u.deg - (pfss_fieldlines['ptth'][i, :] * u.radian).to(u.deg)[:n_valid]
+        radius = ((pfss_fieldlines['ptr'][i, :]) * const.R_sun.to(u.cm))[:n_valid]
+        coord = SkyCoord(lon=lon, lat=lat, radius=radius, frame=frame)
+        fieldlines.append(coord)
+
+    # Magnetic field strengths
+    lon_grid = (pfss_fieldlines['phi'] * u.radian - np.pi * u.radian).to(u.deg).value
+    lat_grid = (np.pi / 2. * u.radian - pfss_fieldlines['theta'] * u.radian).to(u.deg).value
+    radius_grid = pfss_fieldlines['rix'] * const.R_sun.to(u.cm).value
+    B_radius = pfss_fieldlines['br']
+    B_lat = pfss_fieldlines['bth']
+    B_lon = pfss_fieldlines['bph']
+    # Create interpolators
+    B_radius_interpolator = RegularGridInterpolator((radius_grid, lat_grid, lon_grid), B_radius,
+                                                    bounds_error=False, fill_value=None)
+    B_lat_interpolator = RegularGridInterpolator((radius_grid, lat_grid, lon_grid), B_lat,
+                                                 bounds_error=False, fill_value=None)
+    B_lon_interpolator = RegularGridInterpolator((radius_grid, lat_grid, lon_grid), B_lon,
+                                                 bounds_error=False, fill_value=None)
+    # Interpolate values through each line
+    field_strengths = []
+    for f in fieldlines:
+        points = np.stack([f.spherical.distance.to(u.cm).value,
+                           f.spherical.lat.to(u.deg).value,
+                           f.spherical.lon.to(u.deg).value], axis=1)
+        b_r = B_radius_interpolator(points)
+        b_lat = B_lat_interpolator(points)
+        b_lon = B_lon_interpolator(points)
+        field_strengths.append(np.sqrt(b_r**2 + b_lat**2 + b_lon**2) * u.Gauss)
+
+    return [(l, b) for l, b in zip(fieldlines, field_strengths)]
+
+
+@u.quantity_input
+def semi_circular_loop(length: u.cm, observer=None, n_points=1000, offset: u.cm = None):
+    """
+    Generate coordinates for a semi-circular loop
+
+    Parameters
+    ----------
+    length : `~astropy.units.Quantity`
+        Full length of the loop
+    observer : `~astropy.coordinates.SkyCoord`, optional
+        Observer that defines te HCC coordinate system. Effectively, this is the
+        coordinate of the midpoint of the loop.
+    n_points : `int`, optional
+        Number of points in the coordinate
+    offset : `~astropy.units.Quantity`
+        Offset along the HCC y-axis
+    """
+    # Calculate a semi-circular loop
+    s = np.linspace(0, length, n_points)
+    z = length / np.pi * np.sin(np.pi * u.rad * s/length)
+    x = np.sqrt(length**2 / np.pi**2 - z**2)
+    x = np.where(s < length/2, -x, x)
+    # Define origin in HCC coordinates such that the midpoint of the loop
+    # is centered on the origin at the solar surface
+    if observer is None:
+        observer = SkyCoord(lon=0*u.deg,
+                            lat=0*u.deg,
+                            frame=sunpy.coordinates.HeliographicStonyhurst)
+    origin = SkyCoord(x=0*u.km,
+                      y=0*u.km,
+                      z=const.R_sun,
+                      frame=sunpy.coordinates.Heliocentric(observer=observer))
+    # Offset along the y-axis, convenien for creating loop arcades
+    delta_y = 0*u.cm if offset is None else offset
+    return SkyCoord(x=x + origin.x,
+                    y=delta_y*np.ones(x.shape) + origin.y,
+                    z=z + origin.z,
+                    frame=origin.frame)
 
 
 @u.quantity_input
@@ -42,10 +149,9 @@ def synthetic_magnetogram(bottom_left_coord, top_right_coord, shape: u.pixel, ce
     if observer is None:
         observer = sunpy.coordinates.ephemeris.get_earth(time=time_now)
     # Transform to HPC frame
-    bottom_left_coord = bottom_left_coord.transform_to(
-        sunpy.coordinates.Helioprojective(observer=observer))
-    top_right_coord = top_right_coord.transform_to(
-        sunpy.coordinates.Helioprojective(observer=observer))
+    hpc_frame = sunpy.coordinates.Helioprojective(observer=observer, obstime=observer.obstime)
+    bottom_left_coord = bottom_left_coord.transform_to(hpc_frame)
+    top_right_coord = top_right_coord.transform_to(hpc_frame)
     # Setup array
     delta_x = (top_right_coord.Tx - bottom_left_coord.Tx).to(u.arcsec)
     delta_y = (top_right_coord.Ty - bottom_left_coord.Ty).to(u.arcsec)
@@ -55,38 +161,21 @@ def synthetic_magnetogram(bottom_left_coord, top_right_coord, shape: u.pixel, ce
     xphysical, yphysical = np.meshgrid(np.arange(shape[0].value)*shape.unit*dx,
                                        np.arange(shape[1].value)*shape.unit*dy)
     # Add sunspots
-    centers = centers.transform_to(sunpy.coordinates.Helioprojective(observer=observer))
+    centers = centers.transform_to(hpc_frame)
     for c, s, a in zip(centers, sigmas, amplitudes):
         xc_2 = (xphysical - (c.Tx - bottom_left_coord.Tx)).to(u.arcsec).value**2.0
         yc_2 = (yphysical - (c.Ty - bottom_left_coord.Ty)).to(u.arcsec).value**2.0
         data += a.to(u.Gauss).value * np.exp(
             - xc_2 / (2 * s[0].to(u.arcsec).value**2) - yc_2 / (2 * s[1].to(u.arcsec).value**2))
     # Build metadata
-    meta = MetaDict({
-        'telescop': 'synthetic_magnetic_imager',
-        'instrume': 'synthetic_magnetic_imager',
-        'detector': 'synthetic_magnetic_imager',
-        'bunit': 'Gauss',
-        'ctype1': 'HPLN-TAN',
-        'ctype2': 'HPLT-TAN',
-        'hgln_obs': observer.transform_to('heliographic_stonyhurst').lon.to(u.deg).value,
-        'hglt_obs': observer.transform_to('heliographic_stonyhurst').lat.to(u.deg).value,
-        'cunit1': 'arcsec',
-        'cunit2': 'arcsec',
-        'crpix1': (shape[0].value + 1)/2.,
-        'crpix2': (shape[1].value + 1)/2.,
-        'cdelt1': dx.value,
-        'cdelt2': dy.value,
-        'crval1': ((bottom_left_coord.Tx + top_right_coord.Tx)/2.).to(u.arcsec).value,
-        'crval2': ((bottom_left_coord.Ty + top_right_coord.Ty)/2.).to(u.arcsec).value,
-        'dsun_obs': observer.transform_to('heliographic_stonyhurst').radius.to(u.m).value,
-        'dsun_ref': observer.transform_to('heliographic_stonyhurst').radius.to(u.m).value,
-        'rsun_ref': const.R_sun.to(u.m).value,
-        'rsun_obs': ((const.R_sun / observer.transform_to(
-            'heliographic_stonyhurst').radius).decompose() * u.radian).to(u.arcsec).value,
-        't_obs': time_now.iso,
-        'date-obs': time_now.iso,
-    })
+    meta = make_fitswcs_header(
+        data,
+        bottom_left_coord,
+        reference_pixel=(0, 0) * u.pixel,
+        scale=u.Quantity((dx, dy)),
+        instrument='synthetic_magnetic_imager',
+        telescope='synthetic_magnetic_imager',
+    )
     return GenericMap(data, meta)
