+++ synthesizAR/visualize/__init__.py
@@ -2,4 +2,4 @@
 Functions for easily visualizing synthesized observations
 """
 from .aia import *
-from .fieldlines import plot_fieldlines
+from .fieldlines import *
