+++ synthesizAR/visualize/fieldlines.py
@@ -7,12 +7,45 @@ import astropy.units as u
 import astropy.constants as const
 from astropy.time import Time
 from astropy.coordinates import SkyCoord
-from sunpy.util.metadata import MetaDict
-from sunpy.map import GenericMap
+from astropy.visualization import ImageNormalize
+from sunpy.map import GenericMap, make_fitswcs_header
+from sunpy.coordinates import Helioprojective
 
 from synthesizAR.util import is_visible
 
-__all__ = ['plot_fieldlines']
+__all__ = ['plot_fieldlines', 'peek_fieldlines']
+
+
+def peek_fieldlines(magnetogram, fieldlines, **kwargs):
+    """
+    Quick plot of streamlines overplotted on magnetogram
+
+    Parameters
+    ----------
+    magnetogram : `~sunpy.map.Map`
+    fieldlines : `list`
+    """
+    fig = plt.figure(figsize=kwargs.get('figsize', (8, 8)))
+    ax = fig.gca(projection=magnetogram)
+    # Plot map
+    norm = kwargs.get('norm', ImageNormalize(vmin=-1.5e3, vmax=1.5e3))
+    magnetogram.plot(axes=ax, title=False, cmap=kwargs.get('cmap', 'hmimag'), norm=norm)
+    # Grid
+    ax.grid(alpha=0.)
+    magnetogram.draw_grid(axes=ax, grid_spacing=10*u.deg, alpha=0.75, color='k')
+    # Lines
+    line_frequency = kwargs.get('line_frequency', 5)
+    for line in fieldlines[::line_frequency]:
+        coord = line.transform_to(magnetogram.coordinate_frame)
+        # Mask lines behind the solar disk
+        i_visible = np.where(is_visible(coord, magnetogram.observer_coordinate))
+        coord_visible = SkyCoord(Tx=coord.Tx[i_visible], Ty=coord.Ty[i_visible],
+                                 distance=coord.distance[i_visible],
+                                 frame=magnetogram.coordinate_frame)
+        ax.plot_coord(coord_visible, '-', color=kwargs.get('color', 'k'), lw=kwargs.get('lw', 1),
+                      alpha=kwargs.get('alpha', 0.5))
+
+    plt.show()
 
 
 def plot_fieldlines(*coords, observer=None, check_visible=True, **kwargs):
@@ -42,34 +75,19 @@ def plot_fieldlines(*coords, observer=None, check_visible=True, **kwargs):
                             radius=const.au,
                             frame='heliographic_stonyhurst',
                             obstime=Time.now())
-    observer = observer.transform_to('heliographic_stonyhurst')
     # Dummy map
     data = np.ones((10, 10))
-    meta = MetaDict({
-        'ctype1': 'HPLN-TAN',
-        'ctype2': 'HPLT-TAN',
-        'cunit1': 'arcsec',
-        'cunit2': 'arcsec',
-        'crpix1': (data.shape[0] + 1)/2.,
-        'crpix2': (data.shape[1] + 1)/2.,
-        'cdelt1': 1.0,
-        'cdelt2': 1.0,
-        'crval1': 0.0,
-        'crval2': 0.0,
-        'hgln_obs': observer.lon.to(u.deg).value,
-        'hglt_obs': observer.lat.to(u.deg).value,
-        'dsun_obs': observer.radius.to(u.m).value,
-        'dsun_ref': observer.radius.to(u.m).value,
-        'rsun_ref': const.R_sun.to(u.m).value,
-        'rsun_obs': ((const.R_sun/observer.radius).decompose()*u.radian).to(u.arcsec).value,
-        't_obs': observer.obstime.iso,
-        'date-obs': observer.obstime.iso,
-    })
+    coord = SkyCoord(Tx=0*u.arcsec,
+                     Ty=0*u.arcsec,
+                     frame=Helioprojective(observer=observer, obstime=observer.obstime))
+    meta = make_fitswcs_header(data, coord, scale=(1, 1)*u.arcsec/u.pixel,)
     dummy_map = GenericMap(data, meta)
     # Plot coordinates
     fig = kwargs.get('fig', plt.figure(figsize=kwargs.get('figsize', None)))
     ax = kwargs.get('ax', fig.gca(projection=dummy_map))
-    dummy_map.plot(alpha=0, extent=[-1000, 1000, -1000, 1000], title=False, axes=ax)
+    imshow_kwargs = {'alpha': 0, 'extent': [-1000, 1000, -1000, 1000], 'title': False}
+    imshow_kwargs.update(kwargs.get('imshow_kwargs', {}))
+    dummy_map.plot(axes=ax, **imshow_kwargs)
     plot_kwargs = kwargs.get('plot_kwargs', {})
     for coord in coords:
         c = coord.transform_to(dummy_map.coordinate_frame)
